+++ Libplanet/Net/ISwarm.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Libplanet.Net.Messages;
+
+namespace Libplanet.Net
+{
+    internal interface ISwarm
+    {
+        Task<Message> SendMessageWithReplyAsync(
+            BoundPeer peer,
+            Message message,
+            TimeSpan? timeout,
+            CancellationToken cancellationToken);
+
+        void ReplyMessage(Message message);
+    }
+}